% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{bash_job}
\alias{bash_job}
\title{bash job using conda env}
\usage{
bash_job(cmd, conda_env, conda_path = NULL, stdout = TRUE, stderr = FALSE)
}
\arguments{
\item{cmd}{bash command to run (eg., "ls -thlc")}

\item{conda_env}{conda environment name}

\item{conda_path}{conda path; if null, then conda assumed to be in your path}

\item{stdout}{return stdout?}

\item{stderr}{return stderr?}
}
\description{
bash job using conda env
}
